+++ modules/features/account/src/main/java/au/com/shiftyjelly/pocketcasts/account/onboarding/OnboardingFlowComposable.kt
@@ -223,7 +223,7 @@ private fun onLoginToExistingAccount(
         is OnboardingFlow.NewOnboardingAccountUpgrade,
         -> {
             if (subscription == null) {
-                navController.navigate(OldOnboardingFlow.PlusUpgrade.routeWithSource(OnboardingUpgradeSource.LOGIN)) {
+                navController.navigate(OldOnboardingFlow.PlusUpgrade.routeWithSource(source = flow.source, forcePurchase = true)) {
                     // clear backstack after successful login
                     val route = if (FeatureFlag.isEnabled(Feature.NEW_ONBOARDING_ACCOUNT_CREATION)) {
                         OldOnboardingFlow.LOG_IN_OR_SIGN_UP