+++ modules/features/podcasts/src/main/java/au/com/shiftyjelly/pocketcasts/podcasts/view/notifications/EnableNotificationsPromptFragment.kt
@@ -76,7 +76,6 @@ internal class EnableNotificationsPromptFragment : BaseDialogFragment() {
                             .fillMaxSize()
                             .padding(22.dp),
                         onCtaClick = viewModel::handleCtaClick,
-                        onDismissClick = ::finalizeAndDismiss,
                         showNewsletterSection = state.showNewsletterOptIn,
                         isNewsletterSelected = state.subscribedToNewsletter,
                         isNotificationSelected = state.notificationsEnabled,